+++ .php-assignments/mysql/data/ib_buffer_pool
@@ -1,20 +1,49 @@
+0,278
+0,9
+0,283
+0,243
+38,52
+38,51
+38,50
+38,49
+38,48
+38,47
+38,46
+38,45
+38,44
+38,43
+38,42
+38,41
+38,40
+38,39
+38,38
+38,37
+38,36
+38,35
 3,2
-10,227
-10,21
-10,109
-10,17
-10,231
-10,240
-10,13
-10,203
+3,1
+3,3
+10,291
+10,158
+10,273
+10,296
+10,271
+10,125
+10,80
+10,94
 10,2
 10,3
-9,5
+11,11
+11,6
+11,8
+11,12
+9,7
 9,2
 9,3
-8,4
-8,7
 8,8
+8,14
+8,15
+8,9
 8,2
 8,3
 20,3
@@ -23,48 +52,22 @@
 19,3
 19,2
 19,1
-18,1
+17,3
+17,2
+17,1
+16,3
+16,2
+16,1
+15,3
+15,2
+15,1
+14,2
+13,2
 12,3
 12,2
-18,3
 12,1
 11,30
 11,29
 11,28
 11,27
 11,26
-11,25
-11,24
-11,23
-11,22
-11,21
-11,20
-11,19
-11,18
-11,17
-11,16
-14,1
-11,14
-11,13
-11,12
-11,11
-11,9
-11,8
-11,6
-11,5
-11,4
-11,3
-14,3
-13,1
-11,2
-11,1
-10,316
-10,315
-10,314
-10,313
-10,312
-10,311
-10,310
-10,309
-10,308
-10,307
