--- .php-assignments/mysql/data/php_assignments/db.opt
@@ -1,2 +1,0 @@
-default-character-set=utf8mb4
-default-collation=utf8mb4_bin
