+++ assignments/assignment3-3/action/add.php
@@ -1,0 +1,27 @@
+<?php
+    require '../common/connect.php';
+
+    $name = $_POST['name'];
+    $content = $_POST['content'];
+
+    $database_handler = connect();
+
+    $sql = 'INSERT INTO contributions (name, content) VALUES(:name, :content)';
+    try
+    {
+        $statement = $database_handler->prepare($sql);
+        if($statement)
+        {
+            $statement->bindParam(':name', htmlspecialchars($name));
+            $statement->bindParam(':content', htmlspecialchars($content));
+            $statement->execute();
+        }
+    }
+    catch (Throwable $e)
+    {
+        echo $e->getMessage();
+        exit;
+    }
+
+    header('Location: ../completion.php');
+?>
