+++ assignments/assignment3-3/css/style.css
@@ -1,0 +1,33 @@
+.post-text {
+    margin: 0;
+}
+.input-name{
+    margin-bottom: 0.5rem;
+}
+
+.textarea-content{
+    margin-bottom: 0.5rem;
+}
+
+.textarea-content label{
+    display: block;
+}
+
+.textarea-content textarea{
+    display: block;
+}
+
+.post{
+    padding: 1rem;
+    margin: 0 1rem 1rem;
+    border:double;
+    border: double 5px rgb(0, 119, 217);
+}
+
+.post-texts {
+    margin-bottom: 1rem;
+}
+
+.post-texts p {
+    margin: 0.2rem;
+}
