--- assignments/assignment3/completion.php
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html lang="ja">
-
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-</head>
-
-<body>
-    <h1>投稿が完了しました。</h1>
-    <button onclick="location.href='./index.php'">投稿一覧へ戻る</button>
-</body>
-
-</html>
